Objgen save/load and reorganization (#50)
I split out the XML parsing for the object into a separate parser class and consolidated some of the comp_objgen code that would be useful outside of the app. MetaObject byte array now passes over to libcomp upon generating the source file instead of XML so I was able to get rid of the jenky XML hack where it didn't parse in full libcomp side.